# api/openapi.yaml
openapi: 3.0.3
info:
  title: Paid Leave API
  description: An API for managing leave within Massachusetts Paid Family and Medical Leave.
  version: "2022-01-06"
servers:
  - url: /v1
    description: Development server
security:
  - jwt: [] # JWT authentication option
paths:
...
  /employers/add:
    post:
      tags:
        - Employers
      summary: Add an FEIN to the logged in Leave Administrator
      operationId: massgov.pfml.api.employers.employer_add_fein
      requestBody:
        $ref: '#/components/requestBodies/EmployerAddFeinRequestBody'
      responses:
        "201":
          description: FEIN added without errors
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessfulResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/EmployerResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "402":
          description: No verification data for FEIN
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicate employer for user
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Error while adding FEIN
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ErrorResponse"
  /employers/withholding/{employer_id}:
    get:
      summary: Retrieves the last withholding date for the FEIN specified
      operationId: massgov.pfml.api.employers.employer_get_most_recent_withholding_dates
      tags:
        - Employers
      parameters:
        - name: employer_id
          in: path
          schema:
            type: string
            format: uuid
            example: "009fa369-291b-403f-a85a-15e938c26f2f"
          description: "ID of the employer"
          required: true
      responses:
        "200":
          description: A successful request returns a 200 status code.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessfulResponse'
                  - properties:
                      data:
                        $ref: "#/components/schemas/WithholdingResponse"
        "400":
          description: An unsuccessful request returns a 400 status code if one or more errors occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /notifications:
    post:
      security:
        - oauth2: []
      tags:
        - Notifications
      summary: >
        Send a notification that a document is available for a claimant to either the claimant or leave administrator.
      operationId: massgov.pfml.api.notifications.notifications_post
      responses:
        "201":
          description: a successful response returns a 201 status code and asynchronously starts the notification process.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessfulResponse"
                example:
                  meta:
                    method: "POST"
                    resource: "/notifications"
                  status_code: 201
                  message: "success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          description: todo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      requestBody:
        $ref: "#/components/requestBodies/NotificationRequest"
...
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: massgov.pfml.api.authentication.decode_jwt
    oauth2:
      type: oauth2
      x-tokenInfoFunc: massgov.pfml.api.authentication.decode_jwt
      flows:
        clientCredentials:
          tokenUrl: /api/v1/oauth2/token
          scopes: {}
...
  requestBodies:
...
    NotificationRequest:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NotificationRequest"
      required: true
...
    EmployerAddFeinRequestBody:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/EmployerAddFeinRequestBody"
  responses:
...
    BadRequest:
      description: There was a problem with your request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    SuccessfulResponse:
      type: object
      properties:
        status_code:
          type: integer
        message:
          type: string
        meta:
          $ref: "#/components/schemas/Meta"
        data:
          oneOf:
            - type: array
            - type: object
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/ValidationErrorDetail"
      required: ["status_code"]
      additionalProperties: false
    ErrorResponse:
      type: object
      properties:
        status_code:
          type: integer
        message:
          type: string
        meta:
          $ref: "#/components/schemas/Meta"
        data:
          oneOf:
            - type: array
            - type: object
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/ValidationErrorDetail"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ValidationErrorDetail"
      required: ["status_code", "errors"]
      additionalProperties: false
    Meta:
      type: object
      properties:
        resource:
          type: string
        method:
          type: string
        query:
          type: object
        paging:
          type: object
          properties:
            page_offset:
              type: integer
            page_size:
              type: integer
            total_records:
              type: integer
            total_pages:
              type: integer
            order_by:
              type: string
            order_direction:
              type: string
      required: ["resource", "method"]
      additionalProperties: false
    ValidationErrorDetail:
      type: object
      properties:
        type:
          example: "required"
          type: string
        message:
          example: "leave_details.reason is required."
          description: "Somewhat user-friendly description of the error. May contain PII."
          type: string
        rule:
          description: >
            Usage varies depending on the source of the validation error. OpenAPI validation
            errors will set this to the expected value behavior (i.e the max length, or pattern).
          oneOf:
            - type: string
            - type: number
            - type: integer
            - type: boolean
            - type: array
            - type: object
          nullable: true
        field:
          description: "Field path, when the error is associated with a specific field."
          example: "leave_details.reason"
          nullable: true
          type: string
      additionalProperties: false
...
    WithholdingResponse:
      type: object
      properties:
        filing_period:
          format: date
          example: "1970-06-01"
...
    EmployerResponse:
      type: object
      properties:
        employer_id:
          type: string
          format: uuid
          example: "fabd09ed-0ea9-4091-843e-2e37d2c61482"
        employer_fein:
          type: string
          example: "12-3456789"
        employer_dba:
          type: string
          nullable: true
...
    EmployerAddFeinRequestBody:
      type: object
      properties:
        employer_fein:
          $ref: "#/components/schemas/Fein"
          nullable: true
...
    NotificationRecipient:
      type: object
      properties:
        first_name:
          type: string
          example: "Jane"
          description: "Required if and only if recipient_type is Claimant"
        last_name:
          type: string
          example: "Doe"
          description: "Required if and only if recipient_type is Claimant"
        full_name:
          type: string
          example: "Jane Doe"
          description: "Required if and only if recipient_type is Leave Administrator"
        contact_id:
          type: string
          example: "11"
          description: "The contact ID of the recipient, only required if recipient type is a Leave Adminstrator"
        email_address:
          type: string
          format: email
          example: "j.a.doe@gmail.com"
          nullable: false
          minLength: 1
    NotificationClaimant:
      type: object
      properties:
        first_name:
          type: string
          example: "John"
          nullable: false
          minLength: 1
        last_name:
          type: string
          example: "Smith"
          nullable: false
          minLength: 1
        date_of_birth:
          $ref: "#/components/schemas/Date"
          nullable: false
        customer_id:
          type: string
          example: "1234"
          nullable: false
    NotificationRequest:
      type: object
      properties:
        absence_case_id:
          title: Absence Case ID
          description: FINEOS ID for the case the notification is related to.
          type: string
          example: NTN-111-ABS-01
          minLength: 1
        fein:
          type: string
          example: "41-9904348"
        organization_name:
          type: string
          example: "Wayne Enterprises"
        document_type:
          title: Document Type
          description: The type of document that is available to view.
          type: string
          example: Legal Notice
        trigger:
          title: Trigger
          description: The claim state that triggered the notification.
          type: string
          example: claim.approved
        source:
          title: Source
          description: The source of the claim application. Can be either Self-Service or Call Center.
          type: string
          example: Self-Service
          enum:
            - Self-Service
            - Call Center
        recipient_type:
          title: Recipient Type
          description: Who is receiving the notification/what type of person is in the recipients list. Can be either Claimant or Leave Administrator.
          type: string
          example: Leave Administrator
          enum:
            - Claimant
            - Leave Administrator
        recipients:
          type: array
          description: Who to send the notification to.
          items:
            $ref: "#/components/schemas/NotificationRecipient"
        claimant_info:
          $ref: "#/components/schemas/NotificationClaimant"

      required:
        - absence_case_id
        - trigger
        - source
        - recipient_type
        - recipients
        - claimant_info
        - fein
        - organization_name
...
    Fein:
      type: string
      example: "00-0000000"
      pattern: '^\d{2}-\d{7}$'
...
    Date:
      type: string
      example: '1970-01-01'
      format: date
...